Added Pagination Config Option Through Headers.
This will allow for pagination information to be past through the
headers on both the request and response. The request should have a
Range-Unit: items and a Content-Range: 0-4. The 0-4 is the amount of
items per page via the current page.

In addition the response will return the Content-Range: 0-4/10, the 0-4
is the range you requested and the 10 will be the total posts.

This change pairs well with the
angular-paginate-anything library found here:
https://github.com/begriffs/angular-paginate-anything